fix: clear the three warnings GCC 14 raised on the first native UNO Q build
The board's compiler (Debian 13, GCC 14.2 aarch64) is newer than CI's
GCC 12 and MSVC, and flagged three sites neither had:

- SerialMcuBridge::digitalWrite: std::byte{level ? 1u : 0u} narrows
  unsigned to unsigned char in list-initialisation. Cast explicitly.
- mcu::encode: reserve() + push_back() trips GCC 12-14's spurious
  -Wfree-nonheap-object through the inlined _M_realloc_append guard. The
  frame size is known up front, so build it sized and indexed instead -
  no reallocation path either. Wire bytes are unchanged; the framing
  tests and the linked-display round trip prove it.
- runValidation: GCC mis-tracks an optional<string>'s engaged flag when
  copy-assigned into a freshly default-constructed CaseResult and calls
  it maybe-uninitialized. Aggregate-initialise the result instead.

Verified on the board: build clean of all three, platypus_tests 14/14,
scout_validation 21/21. MSVC 19.44 Release: build + ctest still green.

+++ platform/hardware/src/unoq/SerialMcuBridge.cpp
@@ -119,7 +119,7 @@ Status SerialMcuBridge::pinMode(std::uint8_t pin, hal::PinMode mode) {
 }
 
 Status SerialMcuBridge::digitalWrite(std::uint8_t pin, bool level) {
-    const std::byte msg[] = {std::byte{pin}, std::byte{level ? 1u : 0u}};
+    const std::byte msg[] = {std::byte{pin}, std::byte{static_cast<std::uint8_t>(level ? 1 : 0)}};
     return send(mcu::topics::kGpioSet, msg);
 }
